import : store the last revision number on a git note
You need to remeber the last revision you pulled from the mediawiki you
work on, so you will be able to pull the future revisions of this wiki.
The revision number are stored on a note attached to the corresponding
commit in your git repository to do so.
The note can be modified but must begin by the following pattern :
mediawiki_revision: <revision number>