implement the partial cloning of the mediawiki
The command to clone only some pages of a mediawiki
is git clone mediawiki::http://wiki.com##page1##page2...
Split the url and the pages and give the array of the pages
asked. use the API to fetch only the pages of the array and not all
pages like before.

For push and pull the function get_last_remote_revision return
the last revision of the pages tracked on the repository.

The command should be changed when a partial clone will be
implemented in git.