fix git email identity so that commits work
For some reason, Travis doesn't have a hostname that git autodetects
correctly. Even worse, setting it with `git config --global` doesn't
seem to have effect in some tests. Take this job for example:

https://travis-ci.org/Git-Mediawiki/Git-Mediawiki/builds/305944550

The error there is:

    fatal: unable to auto-detect email address (got 'travis@travis-job-fe8c0839-f7af-4467-af53-d4af1d37af69.(none)')

Even though the email is explicitly set in the job itself. For some
reason, in some circumstances (but not all!), autodetection fails. I
suspect it has to do with git running in a deleted directory and
failing to find a local config, because the next job added debugging
information:

https://travis-ci.org/Git-Mediawiki/Git-Mediawiki/builds/305947191

And git config failed with this error:

    fatal: unable to read config file '/home/travis/build/Git-Mediawiki/Git-Mediawiki/t/trash directory.t9361-mw-to-git-push-pull.sh/.gitconfig': No such file or directory

In any case, there's an environment variable (`EMAIL`) used by git to
determine the user's email address that we can use here to fix those
failures.