Completely remove build-time git dependency
Previous build system required the git build tools to be installed to
operate properly. Worse: the test suite required git to be *built* to
operate at all.

We now completely remove that dependency by doing our own "build" or
rather, by removing the need to do a build at all. We simply assume
/usr/bin/perl exists, and that we'll use the GITPERLLIB environment to
specify the path to our Git library during testing.

do not depend on an externally built git and assume it is available