Retry failed mediafile downloads with original $url
When attempting to download a media file, the URL for that file is
extracted from Mediawiki's API response. However, sometimes
that URL does not work for us. One such case is when the URL is
of the form

  https://mediawiki.example.com/path/to/media/file

whereas the original URL we used to contact the server is of the form

  https://user:password@mediawiki.example.com

The URL returned from the API results in 401 Unauthorized, but if we
construct a media file URL corresponding to our original URL:

  https://user:password@mediawiki.example.com/path/to/media/file

we can download the media file successfully.

This patch implements this retry logic automatically: If the URL
supplied by the API fails (for any reason), and there is a difference
between the URL supplied by the API, and the corresponding media file
URL constructed from our original URL, we will retry the request with
our constructed media file URL.

Signed-off-by: Johan Herland <johan@herland.net>